+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 12.2
 - Check subject, course & course period ID are valid for current school & year in Courses.php
 - Fix SQL error: user just deleted current Marking Period, reload Side menu in MarkingPeriods.php
 - Fix SQL error null value in column "marking_period_id" & "teacher_id" in Rollover.php
+- Fix SQL error null value in column "course_period_id" in Rollover.php
 
 Changes in 12.1.3
 -----------------